Fix up whitespace, Add makefile Create key, classify traffic direction and ignore zero payload packets.
Traffic direction classification is basic using port numbers as indicator (assumption is connections will be to port 80 or 443 or some other system port. Other ways this could be done include tracking the connection from the beginning and so observing the direction, knowing which IPs are internall, or looking at handshakes or other indicators.